Move `Quantity[u]` extensions into `Quantity2` trait mixed into companion
Recover the move into the companion by defining the extensions in a
`Quantity2` trait at the package level (outside `object internal`),
then mixing it into `object Quantity`. The trait's body is type-checked
outside the opacity scope, so `Quantity[Measure](1.0)/quantity` resolves
`/` via the `Divisible` typeclass — the resolution that was failing
previously when the extensions lived inside the opacity scope.
