+++ lib/quantitative/src/core/quantitative.Quantity2.scala
@@ -30,16 +30,33 @@
 ┃                                                                                                  ┃
 ┗━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━┛
                                                                                                   */
-package soundness
+package quantitative
 
-export
-  quantitative
-  . { Amount, AmountOfSubstance, Amperes, Atto, Candelas, Celsius, Centi, Current, Deci, Deka,
-      Designation, Distance, Exa, Exbi, Fahrenheit, Femto, Gibi, Giga, Hecto, Kelvins, Kibi, Kilo,
-      Kilograms, Luminosity, Mass, Measure, Mebi, Mega, Metres, Metric, MetricUnit, Micro, Milli,
-      Moles, Nano, NoPrefix, Normalizable, Pebi, Peta, Pico, Principal, Quantifiable, Quantity,
-      Quecto, Quetta, Rankine, Redesignation, Ronna, Ronto, Seconds, Tebi, Temperature, Tera, Time,
-      Units, Yobi, Yocto, Yotta, Zebi, Zepto, Zetta }
+import anticipation.*
+import gossamer.*
+import prepositional.*
+import proscenium.*
+import symbolism.*
 
-package temperatureScales:
-  export quantitative.temperatureScales.{celsius, fahrenheit, kelvin, rankine}
+trait Quantity2:
+  extension [units <: Measure](quantity: Quantity[units])
+    transparent inline def in[units2[power <: Nat] <: Units[power, ?]]: Any =
+      ${quantitative.internal.norm[units, units2]('quantity)}
+
+    transparent inline def invert: Any = Quantity[Measure](1.0)/quantity
+
+
+    inline def normalize[units2 <: Measure](using normalizable: units is Normalizable to units2)
+    :   Quantity[units2] =
+
+      normalizable.normalize(quantity)
+
+
+    inline def sqrt(using root: Quantity[units] is Rootable[2]): root.Result = root.root(quantity)
+    inline def cbrt(using root: Quantity[units] is Rootable[3]): root.Result = root.root(quantity)
+    inline def units: Map[Text, Int] = ${quantitative.internal.collectUnits[units]}
+
+    inline def express(using Decimalizer): Text =
+      t"${quantity.value} ${Quantity.expressUnits(units)}"
+
+    inline def dimension: Text = ${quantitative.internal.describe[units]}
