+++ lib/quantitative/src/core/quantitative_core.scala
@@ -36,28 +36,6 @@ import language.experimental.into
 
 import anticipation.*
 import gossamer.*
-import prepositional.*
-import proscenium.*
-import symbolism.*
-
-extension [units <: Measure](quantity: Quantity[units])
-  transparent inline def in[units2[power <: Nat] <: Units[power, ?]]: Any =
-    ${quantitative.internal.norm[units, units2]('quantity)}
-
-  transparent inline def invert: Any = Quantity[Measure](1.0)/quantity
-
-
-  inline def normalize[units2 <: Measure](using normalizable: units is Normalizable to units2)
-  :   Quantity[units2] =
-
-    normalizable.normalize(quantity)
-
-
-  inline def sqrt(using root: Quantity[units] is Rootable[2]): root.Result = root.root(quantity)
-  inline def cbrt(using root: Quantity[units] is Rootable[3]): root.Result = root.root(quantity)
-  inline def units: Map[Text, Int] = ${quantitative.internal.collectUnits[units]}
-  inline def express(using Decimalizer): Text = t"${quantity.value} ${Quantity.expressUnits(units)}"
-  inline def dimension: Text = ${quantitative.internal.describe[units]}
 
 package temperatureScales:
   given kelvin: TemperatureScale:
